simple-person-sensor/requirements.txt

viam-sdk==0.41.1

typing-extensions
